Minor improvements in "pocket-videos-*" scripts.

+++ pocket-videos-finish-embedding-thumbnails.sh
@@ -4,12 +4,19 @@ set -o errexit
 
 trap "echo FAILED" ERR
 
+thumbnail_png="thumbnail.png"
+if [ -e "$thumbnail_png" ]; then
+	echo >&2 "File with the name already exists, cant't move forward - $thumbnail_png"
+	exit 1
+fi
+
 for thumbnail_orig in "$@"; do
 	echo "$thumbnail_orig"
-	git restore -s HEAD -S -- "$thumbnail_orig"
+	if [ "$(git rev-parse --is-inside-work-tree)" == "true" ]; then
+		git restore -s HEAD -S -- "$thumbnail_orig"
+	fi
 	filename_base="${thumbnail_orig%.*}"
 	filename_video="$filename_base.mp4"
-	thumbnail_png="${filename_base}.png"
 	ffmpeg -i "$thumbnail_orig" "$thumbnail_png"
 	rm "$thumbnail_orig"
 	AtomicParsley "$filename_video" --overWrite --artwork "$thumbnail_png"
